requirements.txt
requests
django-grappelli
django-extensions
django-widget-tweaks
django-countries
django-recaptcha
django-import-export
Django
django-filter
pillow
bibtexparser
gunicorn
psycopg2-binary
model-bakery
django-captcha